@@ -23,7 +23,8 @@ from quacc.wflow_tools.customizers import customize_funcs
 
 if TYPE_CHECKING:
     from collections import UserDict
-    from typing import Any, Callable
+    from collections.abc import Callable
+    from typing import Any
 
     from quacc.types import (
         EspressoPhononDosSchema,
