@@ -258,7 +258,7 @@ def prepare_copy(
     dict
         Dictionary of files to copy.
     """
-    if isinstance(copy_files, (str, Path)):
+    if isinstance(copy_files, str | Path):
         copy_files = [copy_files]
 
     if isinstance(copy_files, list):
